Add DEVICE_Create(), DEVICE_CreateWithAllocator() and DEVICE_GetNext() function.
This commit is related to #1 enhancement.

DEVICE_CreateWithAllocator() allocates resources with a given memory
allocator. DEVICE_Create() is a helper to DEVICE_CreateWithAllocator()
and can be called when a default allocator is enough. Add particular
types:
 - DEVICE_Status to represent valid status codes
 - DEVICE_Handle to represent a device structure
 - DEVICE_MemAllocator to represent memory allocator type
Macro UT_STATUS_EQ() can be used when checking expected and actual
status codes.
Macro COMMON_DIE_IF_NULL() is used to generate error status when given
to function pointer is NULL.